Organized files into 1. duke 2. duke.command 3.duke.exception 4. duke.task packages
