+++ src/main/java/duke/command/ToDoCommand.java
@@ -1,3 +1,8 @@
+package duke.command;
+
+import duke.task.TaskList;
+import duke.Ui;
+
 public class ToDoCommand extends Command {
     private String description;
