+++ src/main/java/duke/task/TaskType.java
@@ -1,3 +1,4 @@
+package duke.task;
 
 public enum TaskType {
     T,
