+++ text-ui-test/EXPECTED.TXT
@@ -4,5 +4,5 @@
 | |_| | |_| |   <  __/
 |____/ \__,_|_|\_\___|
 
-Hello! I'm Duke ┌|*ﾟｏﾟ|┘
+Hello! I'm duke.duke ┌|*ﾟｏﾟ|┘
 What can I do for you?
